sådensguklar

+++ database/migrations/2016_08_17_065345_create_heat_stats_table.php
@@ -17,7 +17,7 @@ class CreateHeatStatsTable extends Migration
             $table->integer('heat_id')->unsigned();
             $table->integer('driver_id')->unsigned();
             $table->integer('vehicle_id')->unsigned()->nullable();
-            $table->integer('time')->nullable();
+            $table->bigInteger('time')->nullable();
             $table->string('distance')->nullable();
             $table->string('fuel_used')->nullable();
             $table->string('kml')->nullable();
